refactor: 당첨 번호 및 보너스 번호 생성 로직 별도의 컨트롤러로 분리
